try torch.profiler
